structure code a bit better
